requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-filter==24.3
celery==5.4.0
redis==5.0.7
phonenumbers==8.13.47
httpx==0.27.2
python-dotenv==1.0.1
uvicorn==0.30.1
gunicorn==22.0.0